pyproject.toml
[project]
name = "aw-connector-example"
version = "0.1.0"
description = "Пример кастомного коннектора для AW BI"
readme = "README.md"
authors = [
    { name = "AW BI", email = "aw@analyticworkspace.ru" }
]
requires-python = ">=3.12"
dependencies = [
    "aiofiles~=24.1",
    "fastapi[standard]~=0.116",
    "polars~=1.32",
    "pyarrow~=21.0",
    "pydantic-settings~=2.10",
    "s3fs~=2025.7",
    "sqlglot~=27.6",
]

[project.scripts]
aw-connector-example = "aw_connector_example:main"

[build-system]
requires = ["uv_build>=0.8.2,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest~=8.4",
]
